Fix missing conversion from int to long

+++ src/pt/up/fe/els2023/table/Value.java
@@ -19,7 +19,7 @@ public class Value {
     }
 
     public static Value of(Integer value) {
-        return new Value(value);
+        return new Value(value.longValue());
     }
 
     public static Value of(Long value) {
